Add copyright header & make sure compilation succeeds on all platforms
